Refactoring: Improve separation og graphql and cypher code.
