+++ xorcery-tbv-neo4j-graphql/src/main/java/com/exoreaction/xorcery/tbv/neo4j/graphqltocypher/TimeVersioningGraphQLToCypherTranslator.java
@@ -1,5 +1,7 @@
-package com.exoreaction.xorcery.tbv.neo4j.graphql;
+package com.exoreaction.xorcery.tbv.neo4j.graphqltocypher;
 
+import com.exoreaction.xorcery.tbv.neo4j.cypher.CypherQueryTransformer;
+import com.exoreaction.xorcery.tbv.neo4j.cypher.TBVCypherConstants;
 import graphql.schema.GraphQLSchema;
 import org.neo4j.graphql.Cypher;
 import org.neo4j.graphql.OptimizedQueryException;
